requirements.txt
tqdm
jupyter
matplotlib
statsmodels
biopython
swalign
esm==0.4.0
